Updating personal fork

+++ monitoring/api/v3/alerts-client/snippets_test.py
@@ -63,8 +63,9 @@ class PochanFixture:
     def __exit__(self, type, value, traceback):
         # Delete the policy and channel we created.
         self.alert_policy_client.delete_alert_policy(self.alert_policy.name)
-        self.notification_channel_client.delete_notification_channel(
-            self.notification_channel.name)
+        if self.notification_channel.name:
+            self.notification_channel_client.delete_notification_channel(
+                self.notification_channel.name)
 
 
 @pytest.fixture(scope='session')
@@ -106,11 +107,20 @@ def test_replace_channels(capsys, pochan):
 
 
 def test_backup_and_restore(capsys, pochan):
-    snippets.backup(pochan.project_name)
+    snippets.backup(pochan.project_name, 'backup.json')
     out, _ = capsys.readouterr()
 
-    snippets.restore(pochan.project_name)
+    snippets.restore(pochan.project_name, 'backup.json')
     out, _ = capsys.readouterr()
     assert "Updated {0}".format(pochan.alert_policy.name) in out
     assert "Updating channel {0}".format(
         pochan.notification_channel.display_name) in out
+
+
+def test_delete_channels(capsys, pochan):
+    notification_channel_id = pochan.notification_channel.name.split('/')[-1]
+    snippets.delete_notification_channels(
+        pochan.project_name, [notification_channel_id], force=True)
+    out, _ = capsys.readouterr()
+    assert "{0} deleted".format(notification_channel_id) in out
+    pochan.notification_channel.name = ''   # So teardown is not tried
